Created an example how to extend the TcpServerNode Node class to create a peer-to-peer network node
An example "skeleton" class is created, named MyPeer2PeerNode, that can be used to start with your own peer-to-peer node for any type of application. Off course, this is the blockchain course, so a blockchain application would be very nice. Furthermore, a specific peer-to-peer network node is created, named AvansNode, to implement a full working node (still under construction) to show how it works and how data is propagated within the network. Eventually, this class will become the example implementation of the blockchain for the course.