Changed approach callback.
Implementation of node could be done by a callback function, receiving all the callbacks from the node and nodeconnections. Or you could derive from the Node and NodeConnection.  To make this more clear, I moved the callback to the methods that should have overridden.  When this is indeed overridden, than the callback is no longer available.